Move unexported methods and structs to internal.go and fix linters
